zwave: make command handling smarter
We have implemented a command queue, which will serialize commands. This
way we can ensure that the events we are getting are (likely) from the
command we are waiting on.

This hasn't been fully tested, and we've seen at least one bug (we are
not publishing some events, for some reason).

Signed-off-by: Joao Eduardo Luis <joao@wipwd.dev>