test: create a small mock server
We sort of need to be able to have a server even on those occasions when
the controller is not physically with us, or when we don't have nodes to
test with.

While this will not allow us to test the gateway, it will still allow us
to test whatever consumes the gateway's data.

Signed-off-by: Joao Eduardo Luis <joao@wipwd.dev>