gateway: deal with some annoyances wrt commands
This may not have (it hasn't) fixed some annoyances about queueing
commands and handling them, but at least seems a decent patch for now.
As long as it works, which we haven't fully tested yet.

Yes, we shouldn't be merging this now. But we sort of need to go for
dinner soon.

Signed-off-by: Joao Eduardo Luis <joao@wipwd.dev>