mqtt: obtain current state
Allows a client to obtain the current state in one go.

We have the feeling this might become unrealistic to perform in a large
enough network. In the future, this might have to be split into
individual messages per-node or something even smaller. One can either
rely on nonces to alert the client these are meant for them, or just use
an alternative topic to publish these and let whomever is listening to
consume them however it is they wish.

Signed-off-by: Joao Eduardo Luis <joao@wipwd.dev>