zwave/config: allow setting config via mqtt
Instead of assuming a configuration file, or a REST api, lets go with
configuring zwave via mqtt. This way we can actually have all zwave
infrastructure over mqtt instead of making a mess with multiple
interfaces.

Configuring mqtt is still an issue, because we can't configure mqtt via
mqtt; we suspect that will actually require a config file, or a very
basic REST api. But that's a problem for future self.

Signed-off-by: Joao Eduardo Luis <joao@wipwd.dev>